Implement rest of CLI (stop, status, etc). Logging cleanup/improvements.
- In the CLI, stop, restart, status, reopen-logs, and ls are now
  implemented.
- The CLI "ls" option is new, and calls perpls to view all the process
  statuses.
- Rename the "gatekeeper-nginx" service in perp to just "nginx".
- Update the logrotate script for the current set of services.
- Switch the perp logging to use svlogd (from the runit library) instead
  of tinylog. It's probably a bit odd to be mixing the logging library
  from runit with perp, but they both function on the same daemontools
  principles. svlogd has more features, though. Longer-term we might use
  the syslog integration, but for now, we're just using it since it
  allows us to disable internal log rotation (tinylog always rotates
  based on file-size). This allows us to more easily manage all of these
  logs with logrotate. While that's somewhat counter to the daemontools
  approach, since we have other log files that can't be managed via the
  process stdout (nginx, trafficserver, elasticsearch), this at least
  allows us to manage all of these log files in a consistent manner.
