+++ bin/api-umbrella-cli
@@ -27,9 +27,15 @@ end
 local cli = require "api-umbrella.cli"
 
 local command = arg[1]
+if not command then
+  command = "help"
+else
+  command = string.gsub(command, "-", "_")
+end
+
 if cli[command] then
   cli[command]()
 else
-  print("Unknown command: " .. tostring(command) .. "\n\n")
+  print("Unknown command: " .. (arg[1] or "") .. "\n\n")
   cli.help()
 end
