+++ templates/etc/elasticsearch/logging.yml.mustache
@@ -6,6 +6,7 @@ logger:
   action: DEBUG
   # reduce the logging for aws, too much is logged under the default INFO
   com.amazonaws: WARN
+  org.apache.http: INFO
 
   # gateway
   #gateway: DEBUG
@@ -33,14 +34,14 @@ appender:
 
   index_search_slow_log_file:
     type: file
-    file: ${path.logs}/elasticsearch-index-search-slowlog.log
+    file: ${path.logs}/elasticsearch/index-search-slowlog.log
     layout:
       type: pattern
       conversionPattern: "[%d{ISO8601}][%-5p][%-25c] %m%n"
 
   index_indexing_slow_log_file:
     type: file
-    file: ${path.logs}/elasticsearch-index-indexing-slowlog.log
+    file: ${path.logs}/elasticsearch/index-indexing-slowlog.log
     layout:
       type: pattern
       conversionPattern: "[%d{ISO8601}][%-5p][%-25c] %m%n"
