--- templates/etc/perp/gatekeeper-nginx/rc.log.mustache
@@ -1,7 +1,0 @@
-#!/bin/sh
-
-if test ${1} = "start"; then
-  exec tinylog -k 50 -s 50000000 -t -z {{log_dir}}/${2}
-fi
-
-exit 0
