+++ templates/etc/perp/nginx/rc.log.mustache
