+++ templates/etc/trafficserver/logs_xml.config
@@ -5,6 +5,6 @@
 
 <LogObject>
   <Format="squid" />
-  <Filename="squid" />
+  <Filename="access.blog" />
   <Mode="binary" />
 </LogObject>
